requirements.txt
google-adk
google-genai
requests
pandas
openpyxl
python-dotenv